ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.1.1
huggingface-hub==0.24.7
pydantic==2.8.2
numpy==1.26.4
pandas==2.2.2
cloudpickle==3.1.1
httpx==0.27.2
openai-whisper==20240930
python-multipart==0.0.12
